fix: revert custom viem helpers

+++ apps/router-api/src/curve-router/curve-router.ts
@@ -3,9 +3,10 @@ import { FastifyBaseLogger } from 'fastify'
 import type { IDict, IRoute, IRouteStep } from '@curvefi/api/lib/interfaces'
 import { PoolTemplate } from '@curvefi/api/lib/pools'
 import { ZERO_ADDRESS as zeroAddress, type Address } from '@primitives/address.utils'
-import { fromWei, toWei, type Decimal } from '@primitives/decimal.utils'
+import { type Decimal } from '@primitives/decimal.utils'
 import { DEFAULT_DECIMALS, notFalsy } from '@primitives/objects.utils'
 import type { RouterRouteResponse, RouteStep, TransactionData } from '@primitives/router.utils'
+import { fromWei, toWei } from '../router.utils'
 import { type RoutesQuery } from '../routes/routes.schemas'
 import { type CurveJS, loadCurve } from './curvejs'
 
@@ -118,7 +119,7 @@ export async function buildCurveRouteResponse(
       router: 'curve',
       routerFeePercentage: '0',
       amountIn: [toWei(fromAmount, fromDecimals)],
-      amountOut: [toWei(toAmount as Decimal, toDecimals)],
+      amountOut: [toWei(toAmount, toDecimals)],
       priceImpact,
       createdAt: Date.now(),
       isStableswapRoute,
